requirements.txt
# VINE-Agent v2: Agentic RAG for Precision Agriculture
# Multi-modal: RAPTOR knowledge + Live sensor context + Drone imagery

# ── Core RAG & LangChain stack ──────────────────────────────────────────────
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
langchain-huggingface>=0.0.6
langchain-openai>=0.1.0
langgraph>=0.1.0

# ── Vector stores & retrieval ────────────────────────────────────────────────
faiss-cpu>=1.7.4
ragatouille>=0.0.9.post2       # ColBERT reranker (optional, graceful fallback)

# ── Embeddings & clustering (RAPTOR) ─────────────────────────────────────────
sentence-transformers>=3.0.0
umap-learn>=0.5.6
scikit-learn>=1.4.0
scipy>=1.12.0
numpy>=1.26.0
pandas>=2.2.0

# ── Drone imagery / geospatial ───────────────────────────────────────────────
rasterio>=1.3.0                # GeoTIFF drone imagery (optional for POC)
geopandas>=0.14.0
shapely>=2.0.0

# ── Data loading ─────────────────────────────────────────────────────────────
openpyxl>=3.1.0                # Excel sensor logs
requests>=2.31.0
datasets>=2.18.0               # HuggingFace agronomic datasets (optional)
pypdf>=4.0.0                   # PDF document loader

# ── Async / HTTP ─────────────────────────────────────────────────────────────
httpx>=0.27.0                  # HTTP client for vLLM server
nest-asyncio>=1.6.0            # Async loop fix in notebooks/some envs

# ── IoT / Sensor ingestion (optional production) ─────────────────────────────
paho-mqtt>=1.6.1               # MQTT client for live sensor feed

# ── Utilities ────────────────────────────────────────────────────────────────
tiktoken>=0.7.0
python-dotenv>=1.0.0
rich>=13.0.0                   # Pretty CLI output
tqdm>=4.66.0

# ── Optional: local LLM / NRP vLLM ──────────────────────────────────────────
# vllm>=0.4.0                  # GPU inference server (uncomment on NRP cluster)
